move math functions from ckb-c-stdlib to lualib
